+++ src/macros.rs
@@ -8,11 +8,11 @@
 /// use cc_traits::{Collection, CollectionRef, covariant_item_ref};
 ///
 /// pub struct MyVec<T>(Vec<T>);
-/// 
+///
 /// impl<T> Collection for MyVec<T> {
 ///   type Item = T;
 /// }
-/// 
+///
 /// impl<T> CollectionRef for MyVec<T> {
 ///   type ItemRef<'a>
 ///   where
@@ -41,9 +41,9 @@ macro_rules! covariant_item_ref {
 /// ```
 /// #![feature(generic_associated_types)]
 /// use cc_traits::{Collection, CollectionMut, covariant_item_mut};
-/// 
+///
 /// pub struct MyVec<T>(Vec<T>);
-/// 
+///
 /// impl<T> Collection for MyVec<T> {
 ///   type Item = T;
 /// }
@@ -76,9 +76,9 @@ macro_rules! covariant_item_mut {
 /// ```
 /// #![feature(generic_associated_types)]
 /// use cc_traits::{Collection, Keyed, KeyedRef, covariant_key_ref};
-/// 
+///
 /// pub struct MyMap<K, V>(std::collections::HashMap<K, V>);
-/// 
+///
 /// impl<K, V> Collection for MyMap<K, V> {
 ///   type Item = V;
 /// }
@@ -86,7 +86,7 @@ macro_rules! covariant_item_mut {
 /// impl<K, V> Keyed for MyMap<K, V> {
 ///   type Key = K;
 /// }
-/// 
+///
 /// impl<K, V> KeyedRef for MyMap<K, V> {
 ///   type KeyRef<'a>
 ///   where