+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.1] - 2021-11-17
+### Added 
+- `GetKeyValue` and `GetKeyValueMut`.
+  Implementations for `HashMap`, `BTreeMap` and `ijson::Object`.
+
 ## [0.6.0] - 2021-10-07
 ### Added 
 - Required `Clone` traits on immutable reference types.